refactor: abstract chat logic into a custom hook for ChatWindow components
